feat: 예약 삭제 시 대기 예약 처리 기능 추가
- 예약 삭제 시 대기 목록이 있을 경우 첫 번째 대기 예약자로 재지정
- 예약 삭제 로직에 대기 목록 처리 로직 추가 및 테스트 코드 작성
- Member와 Reservation 간 양방향 의존성 제거
- JPA 리포지토리에 대기 목록 및 사용자별 예약 조회 메소드 추가

+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -128,10 +128,19 @@ public class ReservationService {
 
     @Transactional
     public void deleteReservationById(final Long id) {
-        reservationRepository.findById(id)
+        Reservation reservation = reservationRepository.findById(id)
             .orElseThrow(() -> new ReservationException("예약을 찾을 수 없습니다."));
 
-        reservationRepository.deleteById(id);
+        List<Waiting> waitings = waitingRepository.findByReservationId(id);
+        if (waitings.isEmpty()) {
+            reservationRepository.deleteById(id);
+            return;
+        }
+
+        Waiting firstWaiting = waitings.getFirst();
+        reservation.reAssignedTo(firstWaiting.getMember());
+
+        waitingRepository.deleteById(firstWaiting.getId());
     }
 
     @Transactional
@@ -164,7 +173,7 @@ public class ReservationService {
         Member member = memberRepository.findById(loginMemberInfo.id())
             .orElseThrow(() -> new BusinessException("멤버를 찾을 수 없습니다."));
 
-        List<Reservation> reservations = member.getReservations();
+        List<Reservation> reservations = reservationRepository.findBy(loginMemberInfo.id());
 
         List<WaitingWithRank> waitingWithRanks = waitingRepository.findByMemberId(loginMemberInfo.id());
